fix: circular dependency issues
