+++ __tests__/unit/lib/utils/cliHelper.test.ts
@@ -9,12 +9,8 @@ import CLIHelper from '../../../../src/utils/cliHelper'
 import { getLatestSupportedVersion } from '../../../../src/metadata/metadataManager'
 import messages from '../../../../src/locales/en'
 import { Work } from '../../../../src/types/work'
-import {
-  readFile,
-  dirExists,
-  fileExists,
-  isGit,
-} from '../../../../src/utils/fsHelper'
+import { isGit } from '../../../../src/utils/fsHelper'
+import { readFile, dirExists, fileExists } from '../../../../src/utils/fsUtils'
 import { format } from 'util'
 
 jest.mock('../../../../src/utils/childProcessUtils', () => {
@@ -39,6 +35,7 @@ jest.mock('../../../../src/utils/repoSetup', () => {
 })
 
 jest.mock('../../../../src/utils/fsHelper')
+jest.mock('../../../../src/utils/fsUtils')
 
 const mockedReadFile = jest.mocked(readFile)
 const mockedDirExists = jest.mocked(dirExists)
